rework the neighborhood scoring system
Gravity-like system was good, but the scoring did not converge to zero nearly
as quickly as it would be great for avoiding the jumps between different
k-neighborhoods.

The new exponential-based weighting gives better results for smaller `k`, thus
increasing the performance a bit.

`boost` parameter was renamed to `smooth` and the values and semantics changed
a bit -- now it is set as a knob in ranges around `-10` to `10`.

Also, bump the version to 1.5.